New finition of types on Book Entity

+++ LibraryBackend/DTO/Books/BookDTO.cs
@@ -10,8 +10,8 @@ namespace LibraryBackend.DTO.Books
         public string LastName { get; set; } = null!;
         public string? AuthorMotherName { get; set; } = null!;
         public string Publisher { get; set; } = null!;
-        public byte ISBN { get; set; }
-        public byte Year { get; set; }
+        public long ISBN { get; set; }
+        public int Year { get; set; }
         public string Collection { get; set; } = null!;
         public byte Copies { get; set; }
     }
